pyproject.toml
[project]
name = "first-steps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.135.1",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.48",
]
